requirements.txt
asgiref==3.3.1
autopep8==1.5.4
boto3==1.16.63
botocore==1.19.63
Django==3.1.5
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
django-storages==1.11.1
gunicorn==20.0.4
jmespath==0.10.0
Pillow==8.1.0
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2020.5
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
whitenoise==5.2.0
